requirements.txt
torch
torchvision
numpy
Pillow
opencv-python
transformers
diffusers
accelerate
safetensors
